backend/requirements.txt
# OptiX Backend Dependencies
fastapi
uvicorn
pydantic
duckdb>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
scipy
sqlalchemy
cachetools
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
python-dotenv
psycopg2-binary
